pyproject.toml
[tool.poetry]
name = "boostcamp-baseline"
version = "0.1.0"
description = ""
authors = ["박성현 <silkcoop@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytorch-lightning = "^2.4.0"
torch = "^2.4.0"
torchvision = "^0.19.0"
omegaconf = "^2.3.0"
numpy = "^2.1.0"
pandas = "^2.2.2"
pillow = "^10.4.0"
requests = "^2.32.3"
dvc = "^3.54.1"
pytest = "^8.3.2"
torchmetrics = "^1.4.1"
wandb = "^0.17.8"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
albumentations = "^1.4.15"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
